Fix chalice deployment bug introduced in #270 * Refactor LocalLambdaClient into seperate file * try/catch import when running from a deployed lambda
